created FileFunctions.java
